Cut down on time complexity by adding an odd number loop. (#46)
In ProtoCompEngineComponent, a conditional if statement was added to the factor computation that uses a slightly different for loop for whether the input is odd or even.
The for loop that was added is for if the value is odd, because it increments by 2 to skip over all even numbers and cut down on time complexity.